part2/csv2kg/pyproject.toml
[project]
name = "graph-fastcampus-part2-csv2kg"
version = "0.1.0"
dependencies = [
    "neo4j~=5.28",
    "python-dotenv>=1.0.0",
    "pandas~=3.0",
    "ipykernel>=6.29.0"
]